web/BE/requirements.txt
blinker==1.7.0
cachetools==5.3.3
click==8.1.7
Flask==3.0.3
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
paho-mqtt==2.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
